PIM-7348: Remove dead completeness code, cleanup
