+++ src/Pim/Component/Catalog/Completeness/CompletenessGeneratorInterface.php
@@ -29,14 +29,6 @@ interface CompletenessGeneratorInterface
     public function generateMissingForChannel(ChannelInterface $channel);
 
     /**
-     * Generate completeness for products given a channel and filters
-     *
-     * @param array $filters
-     * @param ChannelInterface $channel
-     */
-    public function generateMissingForProducts(ChannelInterface $channel, array $filters);
-
-    /**
      * Generate missing completenesses
      */
     public function generateMissing();
